correct controlles,services and auth in micro-users

+++ micro-users/src/main/java/org/cantainercraft/micro/users/service/impl/ProfileServiceImpl.java
@@ -4,9 +4,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.cantainercraft.micro.users.repository.UserRepository;
 import org.cantainercraft.micro.users.service.ProfileService;
+import org.cantainercraft.micro.utilits.exception.ExistResourceException;
+import org.cantainercraft.micro.utilits.exception.NotResourceException;
 import org.cantainercraft.project.entity.users.User;
 import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.cantainercraft.micro.users.convertor.ProfileDTOConvertor;
@@ -32,11 +33,16 @@ public class ProfileServiceImpl implements ProfileService {
     public Profile save (ProfileDTO profileDTO){
         Profile profile = profileDTOConvertor.convertDTOToEntity(profileDTO);
 
+        if(profileRepository.existsByUser(profile.getUser())) throw new ExistResourceException("profile for user is create");
+
         return profileRepository.save(profile);
     }
 
     @Override
     public Profile update(ProfileDTO profileDTO) {
+
+        if(!profileRepository.existsById(profileDTO.getUuid())) throw new NotResourceException("profile is not exist");
+
         Profile profile = profileDTOConvertor.convertDTOToEntity(profileDTO);
         Optional<User> user = userRepository.findById(profileDTO.getUser().getId());
 
@@ -44,6 +50,7 @@ public class ProfileServiceImpl implements ProfileService {
             profile.setUser(e);
             return null;
         });
+
         return profileRepository.save(profile);
     }
 
@@ -67,12 +74,18 @@ public class ProfileServiceImpl implements ProfileService {
     @Override
     @CacheEvict(value = "profiles",key = "#uuid")
     public void deleteById(UUID uuid){
+        if(!profileRepository.existsById(uuid)) throw new NotResourceException("profile is not exist");
+
         profileRepository.deleteById(uuid);
     }
 
     @Override
     @CacheEvict(value = "profiles",allEntries = true)
     public void deleteByUser(Long userId,String email){
+        if(profileRepository.existsByUserIdOrUserEmail(userId,email)){
+            throw new NotResourceException("profile is not exist");
+        }
+
         profileRepository.deleteByUserIdOrUserEmail(userId,email);
     }
 }
